SDAR-50. Use Cincinnati for source Openshift version for integration.
The source Openshift version pulls from Cincinnati for integration
upgrade tests. This is done to support 4.1 -> 4.1 upgrades.

There are several reasons for doing this:
* The current 4.1 nightly is set as default in integration OSD. This
means that the upgrade version is set to the current default in OSD.
This, in turn, falls back on the logic to go to the previous enabled
version to upgrade from.
* The previous version selected is also a nightly, which means the
delta between the two images isn't relevant from a user perspective.
* The 4-stable stream in the release controller includes 4.2, which
means the latest version selected as a source version doesn't make sense
for a 4.1 -> 4.2 upgrade. In other words, there's no way currently to
get a stable stream in the release controller for a major and minor
version, only a major version.
* Finally, 4.1 nightlies are labeled as "4.1.0-0.nightly," which from a
semver perspective is less than 4.1.18. We can't scrub the 4-stable
release controller stream for versions that are less than the nightly
since all of the versions in the 4-stable stream in the release
controller are greater than the nightly.

The end result is that there's no real way of selecting a relevant
lesser version as the source verson of the upgrade to the nightly on
integration. By using Cincinnati, we can use channels that are specific
to the minor version. This allows us to grab the latest stable image as
the source version, then flip over to the release controller to grab the
current nightly version.

This has been done in a way to preserve existing behavior, so other
tests should not change.