SDCICD-40. Use OCM directly for int versions.
We were previously using the release controller for int versions for
initial cluster provisioning. This is a problem because the versions
produced by the release controller are not necessarily present in OSD as
shown in the release controller. By using OCM directly, we can be
certain that the version is present in OCM before provisioning the
cluster.