Extend the poll timeout for CSV on operator upgrade checks.
In the staging environment, the number of hops that can
potentially be made to upgrade an operator up to the latest
version can sometimes exceed the previous polling timeout
of 5 minutes. This increases the timeout to 15 minutes
which should be a generous enough period to allow an operator
to get back to the latest version.

Refs OSD-3141