SDCICD-214. Enable dynamic version selection for upgrades.
Dynamic version selection is now enabled for integration upgrades. Like
dynamic version selection for fresh installs, this allows you to specify
a distance from the current prod default to determine what version of
OCP to upgrade to. This currently only really works for integration, and
it's doubtful this will be expanded unless necessary.

Additionally, the original logic has been modified to support a 0 value
for next release, allowing for the nightly in the current major/minor
version line for the default version in production.