SDCICD-237. Cincinnati aware version selection for upgrades.
Upgrades can now be restricted to only Z releases, not just the latest
version in the OCM versions endpoint.

Stage will now use fast/candidate channels and production will use
stable. Additionally, we will not attempt to select a version if no
edges exist in Cincinnati for the proposed upgrade.