SDCICD-248. Convert gate report into a weather report.
The gate report functionality has now been converted into a weather
report, which will be given to the Service Delivery CI/CD developers
daily covering the last 24 hours in a human readable format. The intent
of the weather report is to describe the failures occurring over the
last 24 hours.

The ability to output Markdown is now supported, and the output of a
weather report will be sorted in order of the job name.

The ability to post a report to a slack webhook is now supported.