SDCICD-280. Remove reliance on OCM.Env in version selection.
All references to "provisioner" have been renamed to "provider" to more
accurately reflect that these objects may be used in situations where no
actual provisioning is occuring.

Version selection no longer relies on OCM.Env for conditional logic,
opting to query the provider object for whether it should use
Cincinnati or the release controller instead. This greatly simplifies
the upgrade logic in pkg/e2e/version.go.

The notion of "production default" has been removed in favor of a
"default version override" coming from the provider. If the default
version override is set, it will be used instead of the default version
found in the version list for a particular provider configuration.

getChannel() has been removed from operator code as well, as it wasn't
really being used and it was an additional area that was relying on
OCM.Env.