SDCICD-280. Remove more references to OCM.Env.
VersionToChannel now uses the provider to determine which Cincinnati
channel to use for upgrades.

OCM.Env has now been removed from metrics. To support this, a new mock
provider has been introduced to ensure that the metrics tests still
behave as expected. The hope is that we'll be able to expand this mock
provider and use it to test much more of osde2e's codebase.

Some general refactoring and expansion of documentation has been done as
part of this effort as well.