OSD-3551. Disable curator operator tests for Openshift version <4.2.0.
The curator operator tests don't work for clusters under version 4.2.0
at least. The curator operator is being removed, but in the meanwhile,
we'll just disable these tests since they don't work for the 4.1.x line.
If they don't work for the 4.2.x line, we'll deal with that when we come
to it.