SDCICD-349. Remove curator operator tests.
The curator operator tests have been removed. The curator operator, in
the coming months, will no longer use the curated source that we
currently test for. Additionally, cluster configurations of various
types on various environments all behave differently. As there's no
deterministic way to determine how this operator should be have, the
quality of the signal generated by this test will be low. Therefore, it
should be removed.