SDCICD-346. Add in version selector unit tests.
Version selector unit tests have been added to ensure that, if we muck
with the logic of OSDe2e version selection in the future, we are able to
verify/validate the existing behavior to ensure we don't break anything.

Upgrade selectors have not been tested due to their reliance on
Cincinnati/the release controller.