add failing tests to capture bug SDCICD-575
This bug appears to occur because the current algorithm for choosing
a random region does not guarantee that it will try all options, nor
does it ensure that the result it uses is actually enabled.

It can both randomly choose to check the same region more than once
(thereby skipping other regions under consideration) and fall through
after trying its fixed quantity of regions without indicating that it
failed to find an enabled one.

Signed-off-by: Chris Waldon <cwaldon@redhat.com>