fix: introduce RetrieveTestResults
I had previously altered RetrieveResults to validate that there
was a present, valid, passing Junit XML in the results, but it
turns out that RetrieveResults is often used for other things.
Changing its behavior broke a lot of stuff throughout the code-
base, so I instead introduced a new method with the additional
validation.

Signed-off-by: Chris Waldon <cwaldon@redhat.com>