fix: accept multi-suite junit xml in e2e harness
This was originally broken because our parsing method only
accepted xml with a single <testsuite> tag, but it's valid
for Junit xml to contain a <testsuites> tag with multiple
child <testsuite> tags.

I first tried to simply use the new parsing library (junit-go)
to do the parse, and that worked fine. However, once parsed
that library cannot re-serialize the xml, so our approach of
loading, transforming, and then saving the XML was no longer
viable.

To work around this, I implemented a custom wrapper type for
ginkgo's report generator that injects the phase names at the
time of initial XML creation. This seems to produce the same
XML and is cleaner to read.

Signed-off-by: Chris Waldon <cwaldon@redhat.com>