deps: stop vendoring dependencies
Vendoring just causes headache when someone forgets to do it, and
it doesn't buy us much. Go has excellent support for package integrity,
so the only thing vendoring accomplishes is giving us an offline copy
of our dependencies. We don't need that, so we can afford to stop
maintaining it.

Signed-off-by: Chris Waldon <cwaldon@redhat.com>