feat: fix watch usage and pass proper client
Previously, everything was failing because I was passing the wrong
instance of rest.Interface into the function (the one powering
the CoreV1() api is not the same as the one for BatchV1()).
That mistake caused the problems I was encountering in the past.

I have also rewritten the usage of Watch() to properly provide
a resourceversion so that we see all changes since the initial
list operation.

Signed-off-by: Chris Waldon <cwaldon@redhat.com>