Handle error when writing metrics to s3 bucket for cluster properties
When the jobName setting is defined, osde2e attempts to write metrics to
s3 bucket. When updating with cluster properties, it was silently
continuing on when the upload to s3 was failing. Added a conditional
check to return when errors occur.

Also includes checking the error returned when uploading to s3 bucket.
When successful, then print the message to the console stating the
upload to message. This helps to clean up the logs.