common/label: initial commit of ginkgo labels
Add a common label package defining a set of Ginkgo labels that can be
used for classifying tests based on what they are doing, what they
support, and how important they are. These labels will be introduced to
tests in following patches to provide a path forward for converting from
mainly using the focus filter to utilizing labels.

Labels provide a great way of grouping tests while not losing the focus
ability currently used today. This should enhance the ability to pick
tests based on the specific criteria when used in combination with the
focus-filter flag.

This list is not all encompassing, new labels can be added as required
and old labels that are no longer relevant should be removed. This list
roughly covers the initial requirements for refactoring a handful of
tests. Any label added should be generic enough to apply to a set of
tests. Focus should be used in conjunction when more granularity past
the category is required.

Signed-off-by: Brady Pratt <bpratt@redhat.com>