Introduce the operators label to the existing operator tests
This change is amending to the existing operator tests describe nodes.
Which allows the suite of tests to be executed using ginkgos label
filter feature over using the focus filter.

Focus filter will continue to work as the string `Suite: operators`
still exists within the test case names. It will eventually be removed
in replacement of the label identifiers.