Enable route verification tests for hypershift clusters
As a customer, they will be accessing the hypershift clusters console in
addition to using the client/API. This change enables the tests
verifying the console url is available/accessible for hypershift
clusters. It does not check for the oauth url as it is not available on
hypershift clusters. Users logged into the cluster are still able to
access oauth endpoint to generate a token for authentication.

Moved the code to get the routes into the ginkgo it statements directly.
It was only being used in one location. Along with some clean up to
existing tests to use the kubernetes-sigs/e2e-framework client.