Destroy cluster after tests with failures when flag is enabled
Currently after test execution finishes, some clean up tasks are
performed and then it checks if test failures occurred, it will return
back a failure. Causing osde2e to end with no cluster clean up
performed.

When the user supplies the option `destroyAfterTest` as true, no matter
what the results are of the tests, the cluster should be destroyed. The
default value for `destroyAfterTest` is true. So clusters should be
deleted for every run.

This commit fixes this issue by moving the destroy cluster block above
the checking if any test failures occurred. Allowing osde2e to properly
destroy the cluster and then analyze the test failures/return.