Remove hypershift elb security group after cluster is deleted
Automates the open bug where the ingress load balancer security group
remains in the aws account after the load balancer is removed. It has an
inbound rule to the vpc used to provision the cluster. Resulting in
being unable to remove the vpc due to the security group/inbound rule.
This will be run each time a hypershift cluster is deleted to remove the
security group and keep the users aws account clean.

It can be removed once the bug is resolved.