Allow osde2e to workaround hypershift admin kubeconfig self cert bug
This commit provides a workaround to osde2e to allow testing to carry on
with the open bug where the cluster certificate does not match the one
the API provides. By default this is not enabled, meaning osde2e will
fail when trying to authenticate with the server. Users will need to
enable this workaround manually by either:
 * Set environment variable `HYPERSHIFT_IGNORE_INVALID_CERT`
 * Set HypershiftIgnoreInvalidCert in a custom config