Fix for rosa to use latest ocp version when not using stable channel
The current implementation will fail trying to create a rosa cluster
when getting the latest version available over the default one
set. Whether using candidate channel or nightly channel, it will append
the channel group to the end of the version string. This behavior works
for traditional OSD cluster installs.

This commit will address the issue with a simple solution to start
(which can later follow with a more concrete implementation based on the
provider for filtering the versions). This change does the following:
 1. Queries only for the versions based on the channel group defined
 2. Attempts to remove channel group that may exist within the version
    string