skip test execution if any failures occur during beforeSuite function
With the more seen intermittent failures of hypershift health checks
failing, osde2e continues on to try to run the tests resulting in
failures.

This change revises the beforeSuite function to be able to skip test
execution based on the error returned. It will not check the type of
error returned and decide what action to take. If the error returned is
a cluster health check error type, it will destroy the cluster if the
user has that flag enabled. The same goes for the other actions
performed, if add ons fail to install, it will delete the cluster based
on the users input.