Fix bug when attempting to delete aws vpc for hypershift clusters
Unable to delete the vpc due to osde2e was not able to locate the
directory on the filesystem where the terraform state file is. In PR #1610
the check to get the report directory from ocm properties was moved.
Which results in trying to get the property after the cluster no longer
exists in ocm.

This commit fixes the issue and moves the check to get ocm property
prior to cluster being destroyed.