Add the ability to write files to a shared directory for later use
This change adds a new config to set a path on the file system for a
shared directory to store files to be used at a later point. The trigger
for this enhancement is when using osde2e in a prow workflow with
following steps that need access to information about the cluster (e.g.
cluster id or kubeconfig). These 2 files are the initial ones that will
be stored in the shared directory.

The shared directory will never be archived from a prow job as long as
the report directory given is the prow artifact directory.

In addition, the temporary directory name has changed to include osde2e
when none is supplied. Making it easier for the osde2e user to
understand what this directory is on their file system.