Skip configure alertmanager operator secret exists tests for nightly build clusters
Configure alertmanager operator secrets (dms-secret and pd-secret) are
not applied to clusters that are deployed using nightly openshift
builds. This is design to avoid paging SRE.

Channel group is fetched from OCM cluster resource as it contains the
channel group used when provisioned. Allowing the test case to be
skipped if a cluster ID is provided and no channel group is defined.