Wait for dedicated-admins-cluster clusterrole to exists for scc tests
Lately we see this test failing to find the dedicated-admins-cluster
clusterrole once the cluster was provisioned. The clusterrole gets
created as part of the helpers (e.g. test setup) but test fails being
unable to find it. Its possible it could be a timing issue where its
being created then test tries to locate it right after.

This commit hopes to resolve this by adding a wait to see if the
clusterrole exists.