verify/pods: ignore pods that are in osde2e prefixed namespaces
The verify pods test is focused on verifying no pods are in a failed
state that are required to mark the cluster is operational "healthy".

If a prior test case ran but had a pod end up in a failed state/the pod
was never properly cleaned up. It causes the verify/pod test to fail.
The prior test should handle its clean up to not pollute following
tests.

This commit ignores any pods that are within an osde2e prefixed
namespace.