helper: introduce retries when attempting to fetch namespace from server
Intermittently tests will fail to retrieve the namespace created for the
test to use. This happens as pre-setup to the test execution. Sometimes
it fails due to the projects api service is unavailable.

This commit adds a retry of 6 attempts (30 seconds total duration) to
attempt to get the namespace from the server. This can help in the event
there was a small hiccup in the network when the pre test setup attempts
to contact the server.