SDCICD-971: add an option to toggle on/off running post upgrade tests
Ginkgo does not support calling RunSpec twice which is what osde2e is
attempting to do for upgrade use cases. Ginkgo does allow to run the
same spec multiple times in a row when invoked from the cli. To remove
the failures in CI jobs/local osde2e runs. This commit adds a new option
to toggle on/off running post upgrade tests. It sets the default to
never run them until a permanent solution is in place.