Add cluster gate agreement when performing cluster upgrades
This change adds support to osde2e to be able to provide cluster
acknowledgements prior to performing a cluster upgrade. Depending on the
cluster upgrade path being performed. Cluster acknowledgement is
performed for both OCM upgrades and local managed upgrade operator
upgrades.

Cluster gate agreement is not performed if the upgrade is a z stream
upgrade, only y stream. In the event a version gate does not exist for
the version to upgrade to, it will still carry on and attempt to perform
the upgrade. To present any errors that are helpful to the user to
debug.

For SDCICD-975.