Add new install version selectors for supporting upgrade workflows
This change brings two new install version selectors. They are added to
support the following upgrade paths:
 * Latest Y-1 to latest Y
 * Latest Z-1 to latest Z

The selectors added are:
 * Install latest Y from delta
 * Install latest Z from delta

They allow the user to define a positive/negative delta to locate the
version to install. These can be used outside of upgrade workflows as
well.

In addition to these version selectors, when querying ocm for openshift
versions, the sorting was switched to be in descending order.