cmd/osde2e: simplify test harness
Funnel a background context down the pipe to the function in which
ginkgo or the test-harness are executed then bump the test harness
launching up a level to simplify the execution. This removes all of the
skipped tests within the osde2e suite and will only show the results of
the tests ran within the harness.

before:

osde2e -> ginkgo -> launch pod -> run tests

after:

osde2e -> launch pod -> run tests

Signed-off-by: Brady Pratt <bpratt@redhat.com>

misc: return err from WriteResults

to allow not using Gomega in a non-test context (harness)

Signed-off-by: Brady Pratt <bpratt@redhat.com>