Do not attempt account role creation when failed to get account roles
This commit resolves a bug in the create/get account role methods. In
the event of getting account roles fails, it should not attempt to create
account roles. With this fix, account roles will only be created if
they do not exist. Any other error wills will be raised back up to the
caller function.