Fix the bug causing rosa cluster oidc config deletion to fail
This commit resolves the bug where after a rosa clusters operator roles
and oidc provider are deleted, it hangs trying to delete the oidc config
if one was auto created when the cluster was created. The issue seemed
to be the rosa oidc config delete command was thinking it was in an
interactive mode, waiting for user prompt. Issue looks to be related to
a global variable being set from prior rosa command calls, resulting it
being set to true, entering interactive mode.

The fix is to add an additional option to explicitly set interactive
mode to false.

In addition to this fix, updated create oidc config to accept the
installer role arn as a parameter over defining a static one.