ROSA provider: perform rosa login prior to deleting cluster/resources
This change performs an ocm login and a rosa login prior to starting the
process to delete the cluster.

This is required when a user of osde2e uses the cleanup command to
delete the cluster (if they did not run osde2e test within the same
container or machine). What happens is rosa is not logged in, resulting
in operator role and oidc provider deletion commands to fail.

When osde2e test is called, prior to creating a cluster. Rosa login
happens when checking for versions. Storing the login authentication
details on the filesystem, resulting in following rosa calls to work.