Test Harness: support multiple reruns of osde2e against the same cluster
Before this change, if a user was to run osde2e test harness, it creates
a namespace for the secrets the tests can access. The namespace
continues to live on the cluster after osde2e finishes. New runs would
fail before running test harness due to the namespace already exists.

This change deletes the secrets namespace as part of the clean up
operations and also attempts to delete the secrets namespace prior to
creating it.