Attempt cluster deletion when failing to create cleanup helper object
Before this change, an end to end run may exhibit the behavior where the
cluster is launched and fails at a certain point where its unable to get
the kubeconfig for it. Which results in failing to construct the cleanup
helper object and cluster deletion is never performed. Leaving cloud
resources up.

This commit will attempt cluster deletion if osde2e fails to construct
the helper object to perform cleanup operations. Resolving the common
case when there is no kubeconfig available. Cluster will be only deleted
based on the skip cluster deletion setting.