fix: increase the delay for expiring clusters
In nightly jobs, with the changes to run as a multi-step job,
must-gather is now ran after the first osde2e process has completed.
This introduced a bug due to osde2e setting an expiration time of 1
minute in the future which is not enough time for the must-gather to
finish.

```
2023/08/25 06:23:49 cluster.go:1241: Successfully set cluster expiry time to Friday, 25 Aug 2023 06:24:48 UTC
```

Increase the time to 10 minutes to be sure that must-gather can run. If
anything fails, the cluster will expire, otherwise osde2e will clean up
the cluster in the post step.

Signed-off-by: Brady Pratt <bpratt@redhat.com>