debug: fix getLastJobID search for job ID
the page has changed since this code previously worked, it is now
dynamically rendered on the client side so the value is not available in
the HTML directly. Parse the last script block on the page for the JSON
object used to render the page. Then look for the second item in the
list, assuming it is the previously ran job.

Yes, this is extremely fragile.

Signed-off-by: Brady Pratt <bpratt@redhat.com>