Reduce sleep time in "inhibits ClusterOperatorDegraded" unit test in `osde2e` inhibitions suite
The `time.Sleep` in this test is currently set to 10 minutes.  This
seems gratuitous and should be reduced, if possible.

This came on our radar because of the ticket
https://issues.redhat.com/browse/OSD-29225, which fixes an issue where
the first unit test defined in the `inhibitions` suite was consistently
failing due to the unit test attempting to do a deep equality check on
slices of different types.

See this document to learn how to test this locally:
https://gitlab.cee.redhat.com/insight-global/docs/-/blob/master/osde2e.md