Test fails because reflect.DeepEqual is comparing different types
All tests in the "should exist" unit test that has an expectedPresent
value of true would fail because reflect.DeepEqual was comparing
disparate types (slice of string and slice of
prometheusModel.LabelName).

The easiest way to fix this was to simply change the type in the
table-based tests anonymous struct to []string. And, as a bonus, this
replaces all of the instances of prometheusModel.LabelName with
[]string, which aligns with the fix needed for the breaking change (see
below).

Here is the fella that caused the bug:
https://github.com/prometheus/alertmanager/commit/34115a78a3400a856021e995a70715f9703f25cb

https://pkg.go.dev/reflect#DeepEqual